% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afsis}
\alias{afsis}
\title{Africa Soil Information Service (AfSIS) XRPD reference library}
\format{
A list of 3 components
\describe{
\item{xrd}{A dataframe of all xrd data (counts only). Column names denote the
reference sample}
\item{tth}{A vector of 2theta intervals of all measurements in the library}
\item{phases}{A dataframe the phase ID's, names and reference intensity
ratios (RIR)}
}
}
\usage{
afsis
}
\description{
A \code{powdRlib} object of 21 pure reference patterns and associated reference
intensity ratios. Data were collected on a Bruker D2
Phaser using Cu K-alpha radiation. All patterns have been normalised to 10,000
counts and reference intensity ratios transformed so that all are relative
to that of corundum, which has been set to a value of 1.0.
}
\keyword{datasets}
