% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afsis_codes}
\alias{afsis_codes}
\title{Original codes for the \code{afsis} reference patterns}
\format{
An 2 column data frame. First column contains the phase IDs from \code{afsis$phase_id}
and the second column the original IDs prior to the inclusion in powdR.
}
\usage{
afsis_codes
}
\description{
A data frame detailing the original codes associated with the \code{afsis} reference
patterns prior to their addition to powdR.
}
\keyword{datasets}
