% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afsis_regroup}
\alias{afsis_regroup}
\title{Regrouping structure for the Africa Soil Information Service (AfSIS) XRPD
reference library}
\format{
A data frame. First column contains the phase IDs present in
\code{afsis$phases$phase_id}. Second column contains the phase names that
constitute the regrouping structure.
}
\usage{
afsis_regroup
}
\description{
A data frame containing an example re-grouping structure for the \code{afsis}
reference library, which results in a slightly coarser description of clay
minerals and Fe/Ti-(hydr)oxides in \code{powdRfps} or \code{powdRafps} objects
when used with \code{regroup()}.
}
\keyword{datasets}
