% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd_align.R
\name{align.optim}
\alias{align.optim}
\title{Maximisisation of correlation}
\usage{
align.optim(a, par, xout, xrd.standard_shorter)
}
\arguments{
\item{a}{a dataframe of the XRPD sample to be aligned}

\item{par}{the shift to be applied (this gets optimised)}

\item{xout}{the tth vector re-approximate the sample to after aligment}

\item{xrd.standard_shorter}{a shortened data frame of an XRPD standard
used to accomodate for clipping of data during alignment}
}
\description{
\code{align.optim} is a function that gets optimised during alignment
when using \code{xrd.align} and \code{multi.xrd.align}. Optimisation
maximises the correlation between a sample and standard.
}
