% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_quant.R
\name{close_quant.powdRafps}
\alias{close_quant.powdRafps}
\title{Close the phase concentration data within a \code{powdRafps} object}
\usage{
\method{close_quant}{powdRafps}(x, ...)
}
\arguments{
\item{x}{A \code{powdRafps} object derived from \code{afps()}.}

\item{...}{other arguments}
}
\value{
a \code{powdRafps} object with components:
\item{tth}{a vector of the 2theta scale of the fitted data}
\item{fitted}{a vector of the fitted XRPD pattern}
\item{measured}{a vector of the original XRPD measurement (aligned and harmonised)}
\item{residuals}{a vector of the residuals (fitted vs measured)}
\item{phases}{a dataframe of the phases used to produce the fitted pattern and their concentrations}
\item{phases_grouped}{the phases dataframe grouped by phase_name and concentrations summed}
\item{obj}{named vector of the objective parameters summarising the quality of the fit}
\item{weighted_pure_patterns}{a dataframe of reference patterns used to produce the fitted pattern.
All patterns have been weighted according to the coefficients used in the fit}
\item{coefficients}{a named vector of coefficients used to produce the fitted pattern}
\item{inputs}{a list of input arguments used in the function call}
}
\description{
\code{close_quant} closes the quantitative data within a \code{powdRafps} object
(derived from \code{afps()}) by ensuring that the composition sums to 100 percent.
}
\examples{
\dontrun{
data(rockjock)
data(rockjock_mixtures)

rockjock_a1 <- afps(lib = rockjock,
                    smpl = rockjock_mixtures$Mix1,
                    std = "CORUNDUM",
                    align = 0.3,
                    lod = 1,
                    std_conc = 20)

sum(rockjock_a1$phases$phase_percent)

rockjock_a1c <- close_quant(rockjock_a1)

sum(rockjock_a1c$phases$phase_percent)

}
}
