% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{D5000}
\alias{D5000}
\title{D5000 library of XRPD patterns}
\format{A list of 3
\describe{
\item{XRD}{A dataframe of all xrd data. Column names denote the reference sample}
\item{TTH}{A vector of 2theta intervals of all measurements in the library}
\item{MINERALS}{A dataframe the mineral ID's, names and RIR's}
}}
\usage{
D5000
}
\description{
The D5000 library contains all the reference patterns and reference intensity
ratios needed for full pattern fitting D5000 data.
}
\keyword{datasets}
