% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{D8}
\alias{D8}
\title{D8 library of XRPD patterns}
\format{A list of 3
\describe{
\item{xrd}{A dataframe of all xrd data. Column names denote the reference sample}
\item{tth}{A vector of 2theta intervals of all measurements in the library}
\item{minerals}{A dataframe the mineral ID's, names and RIR's}
}}
\usage{
D8
}
\description{
The D8 library contains all the reference patterns and reference intensity
ratios needed for full pattern fitting D8 data.
}
\keyword{datasets}
