% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{D8_soil}
\alias{D8_soil}
\title{Soil samples measured by D8}
\format{A list of 2 xy dataframes
\describe{
\item{organic}{An organic soil measured by XRPD}
\item{mineral}{A mineral soil measured by XRPD}
}}
\usage{
D8_soil
}
\description{
Two soil samples measured on a Bruker D8. These are used in examples
}
\keyword{datasets}
