% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_zeros.R
\name{fill.zeros}
\alias{fill.zeros}
\title{Fill zeros in combined quantitative datasets}
\usage{
fill.zeros(quant, lib)
}
\arguments{
\item{quant}{a list of outputs from either \code{fpf} or \code{auto.fpf}
functions}

\item{lib}{a reference library used to extract reference intensity ratios
and hence calculate limits of detection}
}
\value{
a list with components:
\item{filled}{a vector of the 2theta scale of the fitted data}
\item{figures}{a list of figures showing the fitted background and
the internal standard. One figure for each sample in the \code{quant}
input.}
\item{additions}{a vector indicating the total additions for each
sample from the 0 replacement.}
}
\description{
\code{fill.zeros} is designed to be used for datasets where many quantitative
outputs are compiled and resulting in accumulation of 0's and/or NA values.
}
