% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpf.R
\name{fpf}
\alias{fpf}
\title{Full pattern fitting}
\usage{
fpf(s.dir, lib, align_shift, TTH_min, TTH_max, delta_lim, solver, obj.function,
  int_std, fpf_shift, amorphous)
}
\arguments{
\item{s.dir}{Path to the sample to be fitted. The file must be in .xy format (space separated)}

\item{lib}{The XRPD library. A list containing three elements. The first (\code{XRPD}) is a dataframe
containing all pre-measured reference patterns by column. The second (\code{TTH}) is a vector of the
2theta measurement intervals for the reference patterns. Third (\code{MINERALS}) is a data frame
containing the unique ID, mineral name, and reference intensity ratio of each pattern in the library.
The order of \code{XRPD} (by column) and \code{MINERALS} (by row) must be identical.}

\item{align_shift}{The maximum shift that is allowed during initial 2theta alignment (degrees)}

\item{TTH_min}{The minimum value of 2theta used during fitting}

\item{TTH_max}{The maximum value of 2theta used during fitting}

\item{delta_lim}{The tuning parameter used to adjust sensitivity of fit. Must greater than 0 and
less than 100. Lower value = more sensitive. We recommend using 0.01.}

\item{solver}{The optimisation routine to be used. One of \code{c("BFGS", "Nelder-Mead", "CG")}}

\item{obj.function}{The objective function to minimise. One of \code{c("Delta", "R", "Rwp")}.
We recommend Rwp}

\item{int_std}{The mineral name (e.g. "Quartz") to be used as internal standard. Must match a mineral
name in the MINERALS table.}

\item{fpf_shift}{Optional. The maximum shift applied during full pattern fitting.}

\item{amorphous}{Optional. Then name of an amorphous phase to be added to the fitting process. Must
match a name in the MINERALS table.}
}
\description{
\code{fpf} returns estimates of soil mineral concentraitons using full pattern fitting.
}
\details{
This function applies full pattern fitting to an XRPD sample to quantify mineral concentrations.
It requires a library of reference patterns with pre-measured reference intensity ratios.
}
