% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fps.R
\name{fps.powdRlib}
\alias{fps.powdRlib}
\title{\code{fps.powdRlib}}
\usage{
\method{fps}{powdRlib}(lib, smpl, refs, std, tth, align, res, solver, obj, ...)
}
\arguments{
\item{lib}{A \code{powdRlib} object representing the reference library.}

\item{smpl}{A data frame. First column is 2theta, second column is counts}

\item{refs}{A character string of reference pattern ID's from the specified library.
The ID's must match ID's in the \code{lib$phases$phase_id} column.}

\item{std}{The phase ID (e.g. "QUA.1") to be used as internal
standard. Must match an ID provided in the \code{phases} parameter.}

\item{tth}{A vector defining the minimum and maximum 2theta values to be used during
fitting. If not defined, then the full range is used.}

\item{align}{The maximum shift that is allowed during initial 2theta
alignment (degrees). Default = 0.1.}

\item{res}{Optional. An integer used to increase the resolution of the
supplied XRPD data during alignment. Default = 4.}

\item{solver}{The optimisation routine to be used. One of \code{c("BFGS", "Nelder-Mead",
"CG")}. Default = "BFGS".}

\item{obj}{The objective function to minimise. One of \code{c("Delta", "R", "Rwp")}.
Default = \code{"Rwp"}.}

\item{...}{other arguments}
}
\value{
a list with components:
\item{tth}{a vector of the 2theta scale of the fitted data}
\item{fitted}{a vector of the fitted XRPD pattern}
\item{measured}{a vector of the original XRPD measurement}
\item{residuals}{a vector of the Residuals of fitted vs measured}
\item{phases}{a dataframe of the phases used to produce the fitted pattern}
\item{phases_summary}{the phases dataframe grouped by phase_name and summarised (sum)}
\item{rwp}{the Rwp of the fitted vs measured pattern}
\item{weighted_pure_patterns}{a dataframe of reference patterns used to produce the fitted pattern.
All patterns have been weighted according to the coefficients used in the fit}
\item{coefficients}{a named vector of coefficients used to produce the fitted pattern}
}
\description{
\code{fps.powdRlib} returns estimates of phase concentrations using full pattern
fitting of X-ray powder diffraction data.
}
\details{
This function applies full pattern fitting to an XRPD sample to quantify phase
concentrations. It requires a library of reference patterns with
pre-measured reference intensity ratios.
}
\examples{
#Load the minerals library
data(minerals)

# Load the soils data
data(soils)

#Since the reference library is relatively small,
#the whole library can be used at once to get an
#estimate of the phases within each sample.
\dontrun{
fps_sand <-  fps(lib = minerals,
                 smpl = soils$sandstone,
                 tth = c(4.5, 69.5),
                 refs = minerals$phases$phase_id,
                 std = "QUA.1")

fps_lime <- fps(lib = minerals,
                smpl = soils$limestone,
                tth = c(4.5, 69.5),
                refs = minerals$phases$phase_id,
                std = "QUA.1")

fps_granite <- fps(lib = minerals,
                   smpl = soils$granite,
                   tth = c(4.5, 69.5),
                   refs = minerals$phases$phase_id,
                   std = "QUA.1")
}
}
