% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fps_lm.R
\name{fps_lm}
\alias{fps_lm}
\title{Full pattern summation using linear regression}
\usage{
fps_lm(lib, ...)
}
\arguments{
\item{lib}{A \code{powdRlib} object representing the reference library. Created using the
\code{powdRlib} constructor function.}

\item{...}{Other parameters passed to methods e.g. \code{fps_lm.powdRlib}}
}
\value{
a list with components:
\item{tth}{a vector of the 2theta scale of the fitted data}
\item{fitted}{a vector of the fitted XRPD pattern}
\item{measured}{a vector of the original XRPD measurement (aligned)}
\item{residuals}{a vector of the residuals (fitted vs measured)}
\item{phases}{a dataframe of the phases used to produce the fitted pattern}
\item{phases_grouped}{the phases dataframe grouped by phase_name and summed}
\item{weighted_pure_patterns}{a dataframe of reference patterns used to produce the fitted pattern.
All patterns have been weighted according to the coefficients used in the fit}
\item{coefficients}{a named vector of coefficients used to produce the fitted pattern}
\item{inputs}{a list of input arguments used in the function call}
}
\description{
\code{fps_lm} returns a simple fit of a given pattern using linear regression,
where coefficients may be either positive or negative. For more details see
\code{?fps_lm.powdRlib}.
}
\details{
Requires a \code{powdRlib} library of reference patterns. Mineral concentrations
are not quantified and therefore reference intensity ratios are not required.
}
\examples{
#Load the rockjock library
data(rockjock)

# Load the rockjock loadings data
data(rockjock_loadings)

\dontrun{
fps_lm_out <- fps_lm(rockjock,
                     smpl = rockjock_loadings$Dim.1,
                     refs = rockjock$phases$phase_id,
                     std = "QUARTZ",
                     align = 0.3,
                     p = 0.01)

}

}
