% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullpat.R
\name{fullpat}
\alias{fullpat}
\title{Objective measures to be minimised during full pattern fitting}
\usage{
fullpat(par, pure.patterns, sample.pattern, obj, weighting)
}
\arguments{
\item{par}{a named vector of parameters to be optimised}

\item{pure.patterns}{a dataframe of pure patterns (counts only).
The ith column in the dataframe relates to the ith parameter of the
\code{par} vector}

\item{sample.pattern}{a vector of the measured pattern (counts only)
to be fitted.}

\item{obj}{the objective function to be minimised. One of "Delta",
"R" and "Rwp". See page 247 of Bish & Post (1989), Modern Powder
Diffraction.}

\item{weighting}{a weighting pattern (2theta and counts) specifying
which regions to emphasise during full pattern fitting. Use with
caution}
}
\description{
\code{fullpat} is a function that gets optimised during full
pattern fitting when using \code{fpf} and \code{auto.fpf}. Three
objective functions are available for minimisation: Delta, R, and
Rwp. We recommend Rwp.
}
