% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{mc}
\alias{mc}
\title{Mean centre XRPD data}
\usage{
mc(smpl)
}
\arguments{
\item{smpl}{an XRPD dataframe (2theta and counts)}
}
\description{
\code{mc} takes XRPD data (2theta and counts) and
mean centres it by subtracting the mean count intensity from
each measurement interval in the counts vector. This can act
as a form of data normalisation that may be useful when
attempting cluster analysis on XRPD data.
}
\examples{
# Load soil xrd data
data(soils)

#Use the sandstone soil
xrd <- soils$sandstone

xrd$counts <- mc(xrd)
}
