% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_centre.R
\name{mean.centre}
\alias{mean.centre}
\title{Mean centre XRPD data}
\usage{
\method{mean}{centre}(smpl)
}
\arguments{
\item{smpl}{an XRPD dataframe (2theta and counts)}
}
\description{
\code{mean.centre} takes XRPD data (2theta and counts) and
mean centres it by subtracting the mean count intensity from
each measurement interval in the counts vector. This can act
as a form of data normalisation that may be useful when
attempting cluster analysis on XRPD data.
}
\examples{
# Load soil xrd data
data(D8_soil)

#Use second sample in list (mineral soil)
xrd <- D8_soil[[2]]

xrd$counts <- mean.centre(xrd)
}
