% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minerals}
\alias{minerals}
\title{A selection of pure minerals measured on a Bruker D8 arranged into a reference
library.}
\format{A list of 3
\describe{
\item{xrd}{A dataframe of all xrd data (counts only). Column names denote the
reference sample}
\item{tth}{A vector of 2theta intervals of all measurements in the library}
\item{minerals}{A dataframe the mineral ID's, names and RIR's}
}}
\usage{
minerals
}
\description{
The reference library contains a range of measured XRPD data along with reference
intensity ratios needed for full pattern fitting of the \code{soils} data.
}
\keyword{datasets}
