% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minerals_regroup}
\alias{minerals_regroup}
\title{Example regrouping structure for the \code{minerals} data}
\format{
A 2 column data frame.
\describe{First column contains the unique phase IDs of all phases
in the \code{minerals} data. Second column contains the grouping structure for the data
(Non-clay, Clay or Amorphous).}
}
\usage{
minerals_regroup
}
\description{
Example regrouping structure for the \code{minerals} data
}
\keyword{datasets}
