% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minerals_xrd}
\alias{minerals_xrd}
\title{A table of 13 reference patterns and their corresponding two theta scale that can
be combined with a meta data table to create a powdRlib object when using
the \code{create_xrd_lib} function. Use the same layout to create custom
reference libraries.}
\format{A dataframe
\describe{
The first column defines the two theta scale, and remaining columns individual
reference patterns of pure minerals or amorphous phases. Each column title
should be a unique mineral ID
}}
\usage{
minerals_xrd
}
\description{
A table of 13 reference patterns and their corresponding two theta scale that can
be combined with a meta data table to create a powdRlib object when using
the \code{create_xrd_lib} function. Use the same layout to create custom
reference libraries.
}
\keyword{datasets}
