% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_xrpd_align.R
\name{multi_xrpd_align}
\alias{multi_xrpd_align}
\title{Align and harmonise multiple XRPD patterns}
\usage{
multi_xrpd_align(x, standard, xmin, xmax, xshift)
}
\arguments{
\item{x}{a list XRPD dataframes (column 1 = 2theta, column 2 = counts)}

\item{standard}{a dataframe of the chosen standard that each
sample is aligned to (column 1 = 2theta, column 2 = counts)}

\item{xmin}{the minimum 2theta value used during alignment}

\item{xmax}{the maximum 2theta value used during alignment}

\item{xshift}{the maximum (positive and negative) 2theta shift
that is allowed during alignment}
}
\value{
a list of aligned and harmonised XRPD data
}
\description{
\code{multi_xrpd_align} takes a list of XRPD data and aligns them
relative to a chosen standard. It uses an optimisation routine that
computes a suitable linear shift. After all samples have been aligned,
the function harmonises the data to a single 2theta scale.
}
\examples{
# Load soils xrd data
data(soils)

#Load minerals library
data(minerals)

#Create a standard quartz pattern to align to
quartz <- data.frame(tth = minerals$tth,
                     counts = minerals$xrd$QUA.1)

#This function will allow for easier comparison of the peaks
range01 <- function(x){(x-min(x))/(max(x)-min(x))}

#Plot the main quartz peak prior to alignment (scale the counts using rng.nm)
plot(x = soils$sandstone$tth,
     y = range01(soils$sandstone$counts),
     xlim = c(26, 27), type = "l")
lines(x = soils$granite$tth,
      y = range01(soils$granite$counts),
      col = "red")
lines(x = soils$limestone$tth,
      y = range01(soils$limestone$counts),
      col = "green")
lines(x = quartz$tth,
      y = range01(quartz$counts),
      col = "blue")

#align data
aligned <- multi_xrpd_align(soils,
                            standard = quartz,
                            xmin = 10,
                            xmax = 60,
                            xshift = 0.2)
#replot data
plot(x = aligned$sandstone$tth,
     y = range01(aligned$sandstone$counts),
     xlim = c(26, 27), type = "l")
lines(x = aligned$granite$tth,
      y = range01(aligned$granite$counts),
      col = "red")
lines(x = aligned$limestone$tth,
      y = range01(aligned$limestone$counts),
      col = "green")
lines(x = quartz$tth,
      y = range01(quartz$counts),
      col = "blue")

}
