% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.multiXY}
\alias{plot.multiXY}
\title{Plotting a multiXY object}
\usage{
\method{plot}{multiXY}(x, wavelength, interactive, ...)
}
\arguments{
\item{x}{a powdRlib object}

\item{wavelength}{One of "Cu", "Co" or a custom numeric value defining the wavelength
(in Angstroms). Used to compute d-spacings.When "Cu" or "Co" are supplied, wavelengths
of 1.54056 or 1.78897 are used, respectively.}

\item{interactive}{Logical. If TRUE then the output will be an interactive
ggplotly object. If FALSE then the output will be a ggplot object.}

\item{...}{other arguments}
}
\description{
\code{plot.multiXY} is designed to provide easy, adaptable plots
of multiple XRPD patterns.
}
\details{
Plots can be made interactive using the logical \code{interactive} argument.
}
\examples{
# Load the minerals library
data(rockjock_mixtures)
\dontrun{
plot(as_multi_xy(rockjock_mixtures), wavelength = "Cu")
plot(as_multi_xy(rockjock_mixtures), wavelength = "Cu", interactive = TRUE)
}
}
