% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.powdRbkg}
\alias{plot.powdRbkg}
\title{Plotting a powdRbkg object}
\usage{
\method{plot}{powdRbkg}(x, interactive, ...)
}
\arguments{
\item{x}{a powdRlib object}

\item{interactive}{Logical. If TRUE then the output will be an interactive
ggplotly object. If FALSE then the output will be a ggplot object.}

\item{...}{other arguments}
}
\description{
\code{plot.powdRbkg} is designed to provide quick plots to inspect the
fitted backgrounds obtained from \code{bkg}.
}
\details{
The only mandatory argument is x, which must be a powdRbkg object. Plots can
be made interactive using the logical \code{interactive} argument.
}
\examples{
# Load the minerals library
data(minerals)

\dontrun{
plot(minerals, interactive = TRUE)
}
}
