% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powdalign.R
\name{powdalign}
\alias{powdalign}
\title{Align XRD data}
\usage{
powdalign(fd, std, xmin, xmax, nshift, pshift)
}
\arguments{
\item{fd}{A directory containing .xy files of XRPD patterns.}

\item{std}{The file location of the internal standard XRPD pattern.}

\item{xmin}{The minimum 2theta value used during alignment}

\item{xmax}{The maximum 2theta value used during alignment}

\item{nshift}{The negative shift allowance (must be negative)}

\item{pshift}{The positive shift allowance (must be positive)}
}
\value{
A list of aligned XRPD patterns.
}
\description{
\code{powdalign} returns a list of aligned XRPD patterns
}
\details{
This function aligns a list of XRPD sample patterns to a known internal standard
using a linear optimisation.
}
