% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powdRcran.R
\docType{package}
\name{powdRcran}
\alias{powdRcran}
\alias{powdRcran-package}
\title{powdR: A package for quantitative mineralogy.}
\description{
The powdR package provides an implementation of the full pattern summation
approach to quantitative mineralogy from X-ray powder diffraction data.
}
