% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qminerals.R
\name{qminerals}
\alias{qminerals}
\title{Quanity mineral concentrations from optimised parameters and reference
intensity ratios}
\usage{
qminerals(x, xrd.lib)
}
\arguments{
\item{x}{a named vector of optimised parameters from \code{fullpat}}

\item{xrd.lib}{an xrd library containing the reference intensity ratios,
mineral id's and mineral names}
}
\description{
\code{qminerals} returns a two dataframes detailing the mineral concentrations
within a sample analysed by XRPD. \code{min.conc} is used within the \code{fpf}
and \code{auto.fpf} functions.
}
