% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Root mean square error}
\usage{
rmse(pred, obs)
}
\arguments{
\item{pred}{a vector a predicted values}

\item{obs}{a vector of observed values}
}
\description{
\code{rmse} computes the root mean square error of a vector of predicted values
relative to a vector of observed values
}
