% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng_nm.R
\name{rng.nm}
\alias{rng.nm}
\title{Range normalisation}
\usage{
rng.nm(x)
}
\arguments{
\item{x}{a vector to be normalised}
}
\value{
a normalised vector
}
\description{
\code{rng.nm} returns a range normalised vector, where the maximum is normalised
to 1, and the minimum is normalised to 0.
}
