% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_normalisation.R, R/sum_example.R
\name{rngnorm}
\alias{rngnorm}
\alias{rngnorm}
\title{Range normalisation}
\usage{
rngnorm(x)

rngnorm(x)
}
\arguments{
\item{x}{A vector.}

\item{x}{A vector.}
}
\value{
A scaled representation of \code{x}.

A scaled representation of \code{x}.
}
\description{
\code{rngnorm} returns a scaled vector ranging from 0 to 1

\code{rngnorm} returns a scaled vector ranging from 0 to 1
}
\details{
This is a generic function: methods can be defined for it directly or via the
\code{\link{Summary}} group generic. For this to work properly, the arguments
\code{...} should be unnamed, and dispatch is on the first argument.

This is a generic function: methods can be defined for it directly or via the
\code{\link{Summary}} group generic. For this to work properly, the arguments
\code{...} should be unnamed, and dispatch is on the first argument.
}
\examples{
x <- rnorm(100, mean = 0, sd = 100)
rgnorm(x)
x <- rnorm(100, mean = 0, sd = 100)
rgnorm(x)
}
