% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rockjock_loadings}
\alias{rockjock_loadings}
\title{PCA loadings of \code{rockjock_mixtures} data}
\format{
A list of 3 components, each comprised of two columns. Column \code{tth} specifies
the 2theta axis and \code{loading} specifies the count intensities.
}
\usage{
rockjock_loadings
}
\description{
A list containing 3 loadings of Dimensions 1 to 3 from principal components analysis of
the \code{rockjock_mixtures} data.
}
\keyword{datasets}
