% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rockjock_weights}
\alias{rockjock_weights}
\title{Mineral concentrations of the \code{rockjock_mixtures} data}
\format{
An 8 column dataframe, with each row detailing the composition of a sample.
}
\usage{
rockjock_weights
}
\description{
A dataframe summarising the weighed mineral concentrations of the \code{rockjock_mixtures} data,
all in units of weight percent.
}
\references{
Eberl, D.D., 2003. User's guide to RockJock - A program for determining quantitative mineralogy from
powder X-ray diffraction data. Boulder, CA.
}
\author{
Dennis Eberl
}
\keyword{datasets}
