% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bkg.R
\name{run_bkg}
\alias{run_bkg}
\title{Run the background fitting shiny app}
\usage{
run_bkg(...)
}
\arguments{
\item{...}{further arguments to pass to \code{shiny::runApp}}
}
\description{
A wrapper for \code{shiny::runApp} to start the \code{powdR}
background fitting Shiny app.
}
\examples{

\dontrun{

run_powdR()

}

}
