% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_powdR.R
\name{run_powdR}
\alias{run_powdR}
\title{Run the powdR shiny app}
\usage{
run_powdR(...)
}
\arguments{
\item{...}{further arguments to pass to \code{shiny::runApp}}
}
\description{
A wrapper for \code{shiny::runApp} to start the Shiny app for \code{powdR}.
}
\examples{

\dontrun{

run_powdR()

}

}
