% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soils}
\alias{soils}
\title{Example soil XRPD data}
\format{
A multiXY list of 3 XY dataframes (named according to parent material type),
with each XY dataframe containing two columns of:
\describe{
\item{tth}{The 2theta measurement intervals}
\item{counts}{The count intensities}
}
}
\usage{
soils
}
\description{
3 soil samples from different parent materials measured by XRPD (Cu K-alpha radiation)
}
\keyword{datasets}
