% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.powdRlib}
\alias{subset.powdRlib}
\title{##' Subset a powdRlib object
#'
#' \code{subset_powdR} is designed to provide an easy way of
#' subsetting a \code{powdRlib} object by defining the phase ID's
#' that the user wishes to either keep or remove. For more information
#' see \code{?subset_powdR.powdRlib}.
#'
#' @param x a powdRlib object
#' @param ... other arguments
#'
#' @examples
#' #Load the minerals library
#' data(minerals)
#'
#' minerals_keep <- subset_powdR(minerals,
#'                         refs = c("QUA.1", "QUA.2"),
#'                         mode = "keep")
#'
#' minerals_remove <- subset_powdR(minerals,
#'                           refs = c("QUA.1", "QUA.2"),
#'                           mode = "remove")
#' @export
subset_powdR <- function(x, ...) {
  UseMethod("subset_powdR")
}
Subset a powdRlib object}
\usage{
\method{subset}{powdRlib}(x, refs, mode, ...)
}
\arguments{
\item{x}{a powdRlib object}

\item{refs}{a string of the phase ID's of reference patterns to be subset}

\item{mode}{denotes whether the phase ID's defined in the \code{refs} argument are
retained (\code{"keep"}) or removed (\code{"remove"}).}

\item{...}{other arguments}
}
\description{
\code{subset.powdRlib} is designed to provide an easy way of
subsetting a \code{powdRlib} object by defining the phase ID's
that the user wishes to either keep or remove.
}
\examples{
#Load the minerals library
data(minerals)

minerals_keep <- subset(minerals,
                        refs = c("QUA.1", "QUA.2"),
                        mode = "keep")

minerals_remove <- subset(minerals,
                          refs = c("QUA.1", "QUA.2"),
                          mode = "remove")
}
