% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset}
\alias{subset}
\title{Subset a powdRlib object}
\usage{
\method{subset}{powdRlib}(x, ...)
}
\arguments{
\item{x}{a powdRlib object}

\item{...}{other arguments}
}
\description{
\code{subset} is designed to provide an easy way of
subsetting a \code{powdRlib} object by defining the phase ID's
that the user wishes to either keep or remove. For more details
see \code{?subset.powdRlib}.
}
\details{
The only mandatory argument is \code{x}, which must be a powdRlib object.
}
\examples{
#Load the minerals library
data(minerals)

minerals_keep <- subset(minerals,
                        refs = c("QUA.1", "QUA.2"),
                        mode = "keep")

minerals_remove <- subset(minerals,
                          refs = c("QUA.1", "QUA.2"),
                          mode = "remove")
}
