% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_powdR}
\alias{subset_powdR}
\title{Subset a powdRlib object}
\usage{
subset_powdR(lib, ...)
}
\arguments{
\item{lib}{a powdRlib object}

\item{...}{other arguments}
}
\description{
\code{subset_powdR} is designed to provide an easy way of
subsetting a \code{powdRlib} object by defining the phase ID's
that the user wishes to either keep or remove. For more information
see \code{?subset_powdR.powdRlib}.
}
\details{
The only mandatory argument is \code{lib}, which must be a powdRlib object.
}
\examples{
#Load the minerals library
data(minerals)

minerals_keep <- subset_powdR(minerals,
                        refs = c("QUA.1", "QUA.2"),
                        mode = "keep")

minerals_remove <- subset_powdR(minerals,
                          refs = c("QUA.1", "QUA.2"),
                          mode = "remove")
}
