% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_example.R
\name{summ}
\alias{summ}
\title{Summ of vector elements}
\usage{
summ(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
\code{summ} returns the sum of all the values present in its arguments
}
\details{
This is a generic function: methods can be defined for it directly or via the
\code{\link{Summary}} group generic. For this to work properly, the arguments
\code{...} should be unnamed, and dispatch is on the first argument.
}
\examples{
summ(1, 1)
summ(10, 1)
}
