% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_lod.R
\name{tabulate_lod.powdRafps}
\alias{tabulate_lod.powdRafps}
\title{Tabulate limit of detection estimates}
\usage{
\method{tabulate_lod}{powdRafps}(x, type, std_rir, std_lod, ...)
}
\arguments{
\item{x}{a \code{powdRfps} or \code{powdRafps} object.}

\item{type}{one of \code{'all'} or \code{'grouped'}. Specifies whether results
are grouped by phase names. Default = \code{grouped}.}

\item{std_rir}{a numeric value specifying the reference intensity ratio
of a hypothetical internal standard.}

\item{std_lod}{a numeric value specifying an estimated limit of detection
(weight percent) for the hypothetical internal standard.}

\item{...}{other arguments}
}
\value{
a data frame.
}
\description{
\code{tabulate_lod} takes a powdRfps or powdRafps object and
estimates the limit of detection of quantified phases using
a hypothetical limit of detection of an internal standard and
its reference intensity ratio.
}
\examples{

\dontrun{
#Load the minerals library
data(minerals)

Load the soils data
data(soils)

fps_sand <-  fps(lib = minerals,
                 smpl = soils$sandstone,
                 refs = minerals$phases$phase_id,
                 std = "QUA.1",
                 align = 0.2)

#Tabulate lod estimates based on the assumption that
#0.1 \% quartz can be detected, with a RIR of 5.
tabulate_lod(fps_sand,
              std_rir = 5,
              std_lod = 0.1)
}

}
