% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{topsoils}
\alias{topsoils}
\title{NSIS topsoil samples}
\format{A list of 204 dataframes (named according to the sample ID's),
with each dataframe containing two columns of:
\describe{
\item{tth}{The 2theta measurement intervals}
\item{counts}{The count intensities}
}}
\usage{
topsoils
}
\description{
204 samples measured on Bruker D8. Data have had aluminium signal
stripped from them, which was an artefact of the sample holders used
on this instrument.
}
\keyword{datasets}
