% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelength_transform.R
\name{wavelength.transform}
\alias{wavelength.transform}
\title{Adjust 2theta of XRPD data based on monochromatic wavelength}
\usage{
wavelength.transform(x, measured.wavelength, new.wavelength)
}
\arguments{
\item{x}{a vector of 2theta measurement intervals}

\item{measured.wavelength}{the monochromatic X-ray wavelength used
for the original measurement}

\item{new.wavelength}{the monochromatic X-ray wavelength used for
the transformation}
}
\value{
a vector a transformed 2theta intervals
}
\description{
\code{wavelength.transform} returns a new 2theta vector based
on a different monochromatic X-ray wavelength from which the
measurements were made.
}
\examples{
# Load soil xrd data
data(soils)

xrd <- soils$sandstone

#This was measured using Cu radiation (wavelength = 1.54056 Angstroms),
#and will be converted to Co radiation (wavelength = 1.78897 Angstroms)

xrd$tth <- wavelength.transform(xrd$tth, measured.wavelength = 1.54056,
                                new.wavelength = 1.78897)
}
