% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Xpert_NSIS}
\alias{Xpert_NSIS}
\title{Xpert library of XRPD patterns relevant to NSIS soils}
\format{A list of 3
\describe{
\item{xrd}{A dataframe of all xrd data. Column names denote the reference sample}
\item{tth}{A vector of 2theta intervals of all measurements in the library}
\item{minerals}{A dataframe the mineral ID's, names and RIR's}
}}
\usage{
Xpert_NSIS
}
\description{
The Xpert library contains all the reference patterns and reference intensity
ratios needed for full pattern fitting of the NSIS Xpert data.
}
\keyword{datasets}
