% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Xpert_soil}
\alias{Xpert_soil}
\title{Soil samples measured by Xpert}
\format{A list of 2 xy dataframes
\describe{
\item{organic}{An organic soil measured by XRPD}
\item{mineral}{A mineral soil measured by XRPD}
}}
\usage{
Xpert_soil
}
\description{
Two soil samples measured on a Panalytical Xpert. These are used in examples
}
\keyword{datasets}
