% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd_alignment.R
\name{xrd.align}
\alias{xrd.align}
\title{Align XRPD data}
\usage{
xrd.align(xrd.sample, xrd.standard, xmin, xmax, xshift)
}
\arguments{
\item{xrd.sample}{a data frame of the XRPD sample to be aligned. The first column
must be 2theta, and the second must be counts}

\item{xrd.standard}{a  data frame of the XRPD sample to be used as the standard. The
first column must be 2theta, and the second must be counts.}

\item{xmin}{The minimum 2theta value to be used during alignment}

\item{xmax}{the maximum 2theta value to be used during alignment}

\item{xshift}{the maximum 2theta shift that can be applied}
}
\description{
\code{xrd.align} returns a list containing the shift extent (2theta) and a
data frame of the shifted xrpd pattern.
}
\details{
This function applies a linear alignment to an xrpd sample pattern relative to a
chosen standard. The function using brent optimisation to maximise the correlation
between the 2 patterns. Appropriate choice of internal standard is required for
accurate alignment.
}
