% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd_autoid.R
\name{xrd.autoid}
\alias{xrd.autoid}
\title{Automated selection of crystalline phases for quantitative analysis}
\usage{
xrd.autoid(xrd.lib, xrd.sample, delta_lim)
}
\arguments{
\item{xrd.lib}{an dataframe containing XRPD reference patterns (counts only, by column)}

\item{xrd.sample}{a vector of the sample counts}

\item{delta_lim}{a tuning parameter used to adjust the coarseness of the selection.
High values (e.g. 2) will select fewer phases than low values (e.g. 0.01).}
}
\value{
a list with components:
\item{x}{named vector of coefficients derived from qr.solve for each mineral selected}
\item{xrd.lib}{data frame of selected reference patterns}
\item{fit.error}{the trend in total fit error associated with the sequential addition of
phases to the library}
}
\description{
\code{xrd.autoid} is used as part of \code{auto.fpf}. It uses \code{qr.solve} to
rapidly select appropriate phases from a reference library based on the correlation
to the residuals of the fit.
}
\examples{
data(D8_soil)
data(D8_NSIS)

soil <- D8_soil$mineral

auto_select <- xrd.autoid(xrd.lib = D8_NSIS$xrd, xrd.sample = soil$counts, delta_lim = 1)
plot(auto_select$fit.error)
}
