% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd_bin.R
\name{xrd.bin}
\alias{xrd.bin}
\title{Bin XRPD data}
\usage{
xrd.bin(smpl, bin.size)
}
\arguments{
\item{smpl}{a data frame of the XRPD sample to be binned.}

\item{bin.size}{the number of data points to be used in each bin}
}
\value{
a binned XRPD data frame (2theta and counts)
}
\description{
\code{xrd.bin} applies binning to an XRPD dataframe using a specified bin width.
}
\details{
The original values within an XRPD dataframe (2theta and counts) which fall within a
given bin width are replaced by their means for that interval.
}
\examples{
data(D8_soil)

xrd <- D8_soil$mineral
binned <- xrd.bin(smpl = xrd, bin.size = 5)
}
