% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd_bkg.R
\name{xrd.bkg}
\alias{xrd.bkg}
\title{Fit a background to XRPD data}
\usage{
xrd.bkg(tth, counts, width, res)
}
\arguments{
\item{tth}{the 2theta scale}

\item{counts}{the XRPD counts}

\item{width}{the bin width}

\item{res}{the span of the loess polynomial}
}
\value{
a vector of the fitted background
}
\description{
\code{xrd.bkg} fits a background to xrpd data
}
\details{
This function bins the data into chunks and takes the minimum counts
from each chunk. A loess polynomial is then fitted between the minimum
values and the polynomial predicted onto the original 2theta scale.
}
\examples{
data(soils)
xrd <- soils$sandstone
background <- xrd.bkg(tth = xrd$tth, counts = xrd$counts, width = 50, res = 0.1)

plot(xrd, type = "l")
lines(x = xrd$tth, y = background, col = "red")
}
