% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd_fuzzy.R
\name{xrd.fuzzy}
\alias{xrd.fuzzy}
\title{Fuzzy clustering of XRPD data based on first three principal components}
\usage{
xrd.fuzzy(pca, sample.id, nclust, it)
}
\arguments{
\item{pca}{a dataframe containing PC1, PC2 and PC3 scores}

\item{sample.id}{the sample ID's to be assigned to each row of the pca dataframe}

\item{nclust}{the number of clusters to compute}

\item{it}{the maximum number of iterations}
}
\value{
a dataframe containing cluster assignments and membership coefficients
}
\description{
\code{xrd.fuzzy} Applies the Fuzzy c-means clustering algorith from package
e1071 (\code{cmeans}) to the first three principal components of an XRPD
data set.
}
