% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd_lld.R
\name{xrd.lld}
\alias{xrd.lld}
\title{Estimation of lower limits of detection in XRPD data}
\usage{
xrd.lld(x, xrd.sample, xrd.lib, int_std, lld)
}
\arguments{
\item{x}{a named vector of parameters from \code{fullpat}}

\item{xrd.sample}{an XRPD data frame of the sample (2theta, counts)}

\item{xrd.lib}{a library of XRPD reference patterns}

\item{int_std}{the ID of a reference pattern to use as the internal standard}

\item{lld}{a positive tuning parameter used to adjust the lower limit of detection.}
}
\value{
a list with components:
\item{x}{named vector of coefficients derived from full pattern fitting}
\item{xrd.lib}{data frame of reference patterns that are computed to be present in
detectable concentrations}
}
\description{
\code{xrd.lld} uses full patterns and their reference intensity ratios to esimate
the lower limit of detection for crystalline phases.
}
