% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd_pca.R
\name{xrd.pca}
\alias{xrd.pca}
\title{Principal component analysis and pre-treatment of XRPD data}
\usage{
xrd.pca(xrd, align, align.standard, align.xmin, align.xmax, align.xshift, bin,
  bin.width, square.root, normalise)
}
\arguments{
\item{xrd}{a list of XRPD dataframe (2theta and counts)}

\item{align}{logical. If TRUE then alignment will be applied}

\item{align.standard}{required if align = TRUE. Specifies the XRPD
dataframe to be used as the standard for alignment}

\item{align.xmin}{required if align = TRUE. Specifies the minimum 2theta
value used during alignment}

\item{align.xmax}{required if align = TRUE. Specifies the maximum 2theta
value used during alignment}

\item{align.xshift}{required if align = TRUE. Specifies the xmaximum (positive
and negative) 2theta shift that can be used during alignment}

\item{bin}{logical. If TRUE then binning will be applied}

\item{bin.width}{required if bin = TRUE. Specifies the bin width to be used}

\item{square.root}{logical. If TRUE then the count intensities of all XRPD data
will be square root transformed}

\item{normalise}{logical. If TRUE then all XRPD data will be mean centred using
\code{mean.centre}}
}
\value{
a list with components:
\item{pca}{a dataframe of PCA scores}
\item{pca_variance}{a vector of the cumulative proportion of variance explained
by each principal component}
\item{xrd}{a list of the treated XRPD data}
}
\description{
\code{xrd.pca} allows for principal components analysis to be applied to
XRPD subjected to various pre-treatment options.
}
\details{
Four pre-treament options are available. These include peak alignment using
\code{multi.xrd.align}, binning using the \code{xrd.bin}, square root transform,
and normalisation using \code{mean.centre}.
}
