% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autofpf.R
\name{auto.fpf}
\alias{auto.fpf}
\title{Automated full pattern fitting}
\usage{
auto.fpf(smpl, lib, tth, std, amorphous, coarse, align, solver, obj, shift,
  weighting, lld, amorphous_lld)
}
\arguments{
\item{smpl}{A data frame. First column is 2theta, second column is counts}

\item{lib}{The XRPD library. A list containing three elements. The first (\code{XRPD}) is a dataframe
containing all pre-measured reference patterns by column. The second (\code{tth}) is a vector of the
2theta measurement intervals for the reference patterns. Third (\code{MINERALS}) is a data frame
containing the unique ID, mineral name, and reference intensity ratio of each pattern in the library.
The order of \code{XRPD} (by column) and \code{MINERALS} (by row) must be identical.}

\item{tth}{A vector defining the minimum and maximum 2theta values to be used during
fitting}

\item{std}{The mineral ID (e.g. "Qzt.662070.Strath.12Mins.P") to be used as internal standard. Must match a mineral
name in the MINERALS table.}

\item{amorphous}{Optional. Then name of an amorphous phase to be added to the fitting process. Must
match an ID in the MINERALS table.}

\item{coarse}{The tuning parameter used to adjust coarseness of fit. Must be greater than 0 and
less than 100. Low values are more sensitive to minor phases. Default = 0.1.}

\item{align}{The maximum shift that is allowed during initial 2theta alignment (degrees). Default = 0.1.}

\item{solver}{The optimisation routine to be used. One of \code{c("BFGS", "Nelder-Mead", "CG")}. Default = \code{"BFGS"}.}

\item{obj}{The objective function to minimise. One of \code{c("Delta", "R", "Rwp")}. Default = \code{"Rwp"}.}

\item{shift}{Optional. The maximum shift applied during full pattern fitting. Default = 0.05.}

\item{weighting}{Optional. A two column dataframe. First column contains 2theta axis on same scale as that
of the XRD library. Second column contains the weighting of each 2theta variable. If not provided, all
variables are given a weighting of 1.}

\item{lld}{Optional parameter used to tune the lower limit of dection computation.
Must be between 0 and 1. Default = 0.3.}

\item{amorphous_lld}{Optional parameter used to exclude amorphous phases if they are below the
specified \code{amorphous_lld} (percent). Must be between 0 and 100. Default = 0.}
}
\value{
a list with components:
\item{tth}{a vector of the 2theta scale of the fitted data}
\item{fitted}{a vector of the fitted XRPD pattern}
\item{measured}{a vector of the original XRPD measurement}
\item{residuals}{a vector of the Residuals of fitted vs measured}
\item{minerals}{a dataframe of the minerals used to produce fitted}
\item{minerals_summary}{the MINERALS dataframe grouped by minerals and summarised (mean)}
\item{rwp}{the Rwp of the fitted vs measured pattern}
\item{weighted_pure_patterns}{a dataframe of reference patterns used to produce the fitted
pattern. All patterns have been weighted according to the coefficients used in the fit}
\item{coefficients}{a named vector of coefficients used to produce the fitted pattern}
}
\description{
\code{auto.fpf} returns estimates of soil mineral concentraitons using full pattern
 fitting. This automated version attempts to extract the most appropriate phases
 from the library.
}
\details{
This function applies full pattern fitting to an XRPD sample to quantify mineral concentrations.
It requires a library of reference patterns with pre-measured reference intensity ratios.
}
\examples{
# Load the minerals library
data(minerals)

# Load the soils data
data(soils)

# automated without any amorphous phases
\dontrun{
fpf_out <- auto.fpf(smpl = soils$sandstone,
                    lib = minerals,
                    tth = c(4.5, 69.5),
                    std = "QUARTZ.STRATH.P.1142250")
}
#automated with an amorphous phase (organic matter)
\dontrun{
fpf_out_org <- auto.fpf(smpl = soils$sandstone,
                    lib = minerals,
                    tth = c(4.5, 69.5),
                    std = "QUARTZ.STRATH.P.1142250",
                    amorphous = "ORGANIC.bez.CRAIGLICHT.668085.DEEP")
}
# An example of using weighting
weighting <- data.frame(tth = minerals$tth,
                        counts = rep(1, length(minerals$tth)))

# Make all values below tth = 5 have a weighting of 0
weighting$counts[which(weighting$tth <= 5)] <- 0

# Make all values between tth = 26 and 27 have a weighting of 2
weighting$counts[which(weighting$tth >= 26 & weighting$tth <= 27)] <- 10
}
