% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_spacing.R
\name{d.spacing}
\alias{d.spacing}
\title{Calculate d-spacing}
\usage{
d.spacing(x, wavelength)
}
\arguments{
\item{x}{a vector of 2theta measurement intervals}

\item{wavelength}{the monochromatic X-ray wavelength}
}
\description{
\code{d.spacing} returns a vector of d-spacing computed from the supplied
2theta vector and monochromatic wavelength.
}
\examples{
# Load soil xrd data
data(soils)

#Use sandstone sample
xrd <- soils$sandstone

xrd$d <- d.spacing(x = xrd$tth, wavelength = 1.5406)
}
