% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpf_plot.R
\name{fpf.plot}
\alias{fpf.plot}
\title{Interactive full pattern fitting plot}
\usage{
fpf.plot(x, d, wavelength)
}
\arguments{
\item{x}{An \code{fpf} or \code{auto.fpf} output to be plotted}

\item{d}{(optional) logical parameter specifying whether to compute d spacing.
Default = FALSE}

\item{wavelength}{(required only with \code{d = TRUE}) the wavelength samples were measured at
(in Angstroms). Changes x-axis to d-spacing.}
}
\description{
\code{fpf.plot} returns a ggplotly object.
}
\details{
This function is for use in combination with the output from \code{fpf} or \code{auto.fpf}.
}
\examples{
# Load the Xpert library
data(Xpert)

# Load the Xpert soil data to use in example
data(Xpert_soil)

# define the phases to include in the fit
xrd_phases <- c("Qzt.662070.Strath.12Mins.P", "Qzt.662074.Qua.10.P", "X996730.QUA.11.P")

# without organic
\dontrun{
fpf_out <-  fpf(smpl = Xpert_soil$mineral,
               lib = Xpert,
               int.std = "Qzt.662070.Strath.12Mins.P",
               phases = xrd_phases,
               tth = c(4.5, 69.5))
}

# Plot data on 2theta scale
\dontrun{
fpf.plot(fpf_out)
}

# Plot data d-spacing scale
\dontrun{
fpf.plot(fpf_out, d = TRUE, wavelength = 1.54060)
}
}
