% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpf_plot.R
\name{fpf.plot}
\alias{fpf.plot}
\title{Interactive full pattern fitting plot}
\usage{
fpf.plot(x, d, wavelength)
}
\arguments{
\item{x}{An \code{fpf} or \code{auto.fpf} output to be plotted}

\item{d}{(optional) logical parameter specifying whether to compute d spacing.
Default = FALSE}

\item{wavelength}{(required only with \code{d = TRUE}) the wavelength samples were measured at
(in Angstroms). Changes x-axis to d-spacing.}
}
\description{
\code{fpf.plot} returns a ggplotly object.
}
\details{
This function is for use in combination with the output from \code{fpf} or \code{auto.fpf}.
}
\examples{
# Load the minerals library
data(minerals)

# Load the soils data
data(soils)

# define the phases to include in the fit
c_phases <- c("FEL", "KAO",
"OLI", "QUA.1", "QUA.2")

\dontrun{
fpf_out <-  fpf(smpl = soils$sandstone,
               lib = minerals,
               tth = c(4.5, 69.5),
               crystalline = c_phases,
               std = "QUA.1",
               amorphous = "ORG")
}

# Plot data on 2theta scale
\dontrun{
fpf.plot(fpf_out)
}

# Plot data d-spacing scale
\dontrun{
fpf.plot(fpf_out, d = TRUE, wavelength = 1.54060)
}
}
