% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpf_plot.R
\name{fpf.plot}
\alias{fpf.plot}
\title{Interactive full pattern fitting plot}
\usage{
fpf.plot(x, plot.width, plot.height)
}
\arguments{
\item{x}{An \code{fpf} output to be plotted}

\item{plot.width}{Numeric. Width of the plot output (pixels)}

\item{plot.height}{Numeric. Height of the plot output (pixels)}
}
\description{
\code{fpf.plot} returns a ggplotly interactive graph.
}
\details{
This function is for use in combination with the output from \code{fpf}.
}
