% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpf.R
\name{fpf}
\alias{fpf}
\title{Full pattern fitting}
\usage{
fpf(smpl, lib, phases, std, amorphous, tth, align = 0.1, solver = "BFGS",
  obj = "Rwp", shift = 0.05, weighting)
}
\arguments{
\item{smpl}{A data frame. First column is 2theta, second column is counts}

\item{lib}{The XRPD library. A list containing three elements. The first (\code{XRPD})
is a dataframe containing all pre-measured reference patterns by column. The second
(\code{TTH}) is a vector of the 2theta measurement intervals for the reference patterns.
Third (\code{MINERALS}) is a data frame containing the unique ID, mineral name, and
reference intensity ratio of each pattern in the library. The order of \code{XRPD}
(by column) and \code{MINERALS} (by row) must be identical.}

\item{phases}{A string of mineral ID's used to subset the library.}

\item{std}{The mineral ID (e.g. "Qzt.662070.Strath.12Mins.P") to be used as internal
standard. Must match a mineral name in the MINERALS table.}

\item{amorphous}{Optional. The ID of an amorphous phase (e.g. "ORGANIC.337666")to be
added to the fitting process. Must match an ID provided in the \code{phases} parameter.
Only 1 amorphous phase can be used.}

\item{tth}{A vector defining the minimum and maximum 2theta values to be used during
fitting}

\item{align}{The maximum shift that is allowed during initial 2theta alignment (degrees). Default = 0.1}

\item{solver}{The optimisation routine to be used. One of \code{c("BFGS", "Nelder-Mead",
"CG")}. Default = "BFGS".}

\item{obj}{The objective function to minimise. One of \code{c("Delta", "R", "Rwp")}.
Default = "Rwp". Default = Rwp}

\item{shift}{Optional. The maximum shift applied during full pattern fitting.
Default = 0.05}

\item{weighting}{Optional. A column dataframe. First column contains 2theta axis on same scale as that
of the XRD library. Second column contains the weighting of each 2theta variable. If not provided, all variables
are given a weighting of 1.}
}
\value{
a list with components:
\item{TTH}{A vector of the 2theta scale of the fitted data}
\item{FITTED}{A vector of the fitted XRPD pattern}
\item{MEASURED}{A vector of the original XRPD measurement}
\item{RESIDUALS}{A vector of the Residuals of FITTED vs MEASURED}
\item{MINERALS}{A dataframe of the minerals used to produced FITTED}
\item{MINERALS_SUMMARY}{The MINERALS dataframe grouped by minerals and summarised (mean)}
\item{R}{The Rwp of the FITTED vs MEASURED pattern}
\item{WEIGHTED_PURE_PATTERNS}{A dataframe of reference patterns used to produce FITTED.
All patterns have been weighted according to the coefficients used in the fit}
\item{COEFFICIENTS}{A named vector of coefficients used to produce FITTED}
}
\description{
\code{fpf} returns estimates of soil mineral concentraitons using full pattern fitting.
}
\details{
This function applies full pattern fitting to an XRPD sample to quantify mineral
concentrations.
It requires a library of reference patterns with pre-measured reference intensity ratios.
}
\examples{
# Load the Xpert library
data(Xpert)

# Load the Xpert soil data to use in example
data(Xpert_soil)

# define the phases to include in the fit
xrd_phases <- c("Qzt.662070.Strath.12Mins.P", "Qzt.662074.Qua.10.P", "X996730.QUA.11.P")

# without organic
#not run
#fpf_out <-  fpf(smpl = Xpert_soil$mineral,
#               lib = Xpert,
#               tth = c(3.5, 69.5),
#               phases = xrd_phases,
#               std = "Qzt.662070.Strath.12Mins.P")

# Try fitting the same sample, but including an amorphous phase (organic matter)
xrd_phases_org <- c("Qzt.662070.Strath.12Mins.P", "Qzt.662074.Qua.10.P",
                   "X996730.QUA.11.P", "ORGANIC.337666")
# not run
# fpf_out_org <-  fpf(smpl = Xpert_soil$mineral,
#                    lib = Xpert,
#                    tth = c(3.5, 69.5),
#                    phases = xrd_phases_org,
#                    std = "Qzt.662070.Strath.12Mins.P",
#                    amorphous = "ORGANIC.337666")

# An example of using weighting
weighting <- data.frame(TTH = Xpert$TTH,
                        COUNTS = rep(1, length(Xpert$TTH)))

# Make all values below TTH = 5 have a weighting of 0
weighting$COUNTS[which(weighting$TTH <= 5)] <- 0

# Make all values between TTH = 26 and 27 have a weighting of 2
weighting$COUNTS[which(weighting$TTH >= 26 & weighting$TTH <= 27)] <- 10
}
