% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fps.R
\name{fps}
\alias{fps}
\title{Full pattern summation}
\usage{
fps(lib, ...)
}
\arguments{
\item{lib}{A \code{powdRlib} object representing the reference library.}

\item{...}{Other parameters passed to methods e.g. \code{fps.powdRlib}}
}
\value{
a list with components:
\item{tth}{a vector of the 2theta scale of the fitted data}
\item{fitted}{a vector of the fitted XRPD pattern}
\item{measured}{a vector of the original XRPD measurement}
\item{residuals}{a vector of the Residuals of fitted vs measured}
\item{phases}{a dataframe of the phases used to produce the fitted pattern}
\item{phases_summary}{the phases dataframe grouped by phase and summarised (sum)}
\item{rwp}{the Rwp of the fitted vs measured pattern}
\item{weighted_pure_patterns}{a dataframe of reference patterns used to produce the fitted pattern.
All patterns have been weighted according to the coefficients used in the fit}
\item{coefficients}{a named vector of coefficients used to produce the fitted pattern}
}
\description{
\code{fps} returns estimates of phase concentrations using full pattern
fitting of X-ray powder diffraction data. For more details see \code{?fps.powdRlib}.
}
\details{
This function applies full pattern fitting to an XRPD sample to quantify phase
concentrations. It requires a library of reference patterns with
pre-measured reference intensity ratios.
}
\examples{
#Load the minerals library
data(minerals)

# Load the soils data
data(soils)

#Since the reference library is relatively small,
#the whole library can be used at once to get an
#estimate of the phases within each sample.
\dontrun{
fps_sand <-  fps(lib = minerals,
                 smpl = soils$sandstone,
                 tth = c(4.5, 69.5),
                 refs = minerals$phases$phase_id,
                 std = "QUA.1")

fps_lime <- fps(lib = minerals,
                smpl = soils$limestone,
                tth = c(4.5, 69.5),
                refs = minerals$phases$phase_id,
                std = "QUA.1")

fps_granite <- fps(lib = minerals,
                   smpl = soils$granite,
                   tth = c(4.5, 69.5),
                   refs = minerals$phases$phase_id,
                   std = "QUA.1")
}
}
