% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fps.R
\name{fps}
\alias{fps}
\title{Full pattern summation}
\usage{
fps(
  lib,
  smpl,
  harmonise,
  solver,
  obj,
  refs,
  std,
  force,
  std_conc,
  omit_std,
  normalise,
  closed,
  tth_align,
  align,
  manual_align,
  tth_fps,
  shift,
  remove_trace,
  weighting,
  ...
)
}
\arguments{
\item{lib}{A \code{powdRlib} object representing the reference library. Created using the
\code{powdRlib} constructor function.}

\item{smpl}{A data frame. First column is 2theta, second column is counts}

\item{harmonise}{logical parameter defining whether to harmonise the \code{lib} and \code{smpl}.
Default = \code{TRUE}. When \code{TRUE} the function will harmonise the \code{lib} and
\code{smpl} to the intersecting 2theta range at the coarsest resolution available using
natural splines.}

\item{solver}{The optimisation routine to be used. One of \code{"BFGS", "Nelder-Mead",
"CG", "NNLS"}. Default = \code{"BFGS"}.}

\item{obj}{The objective function to minimise when "BFGS", "Nelder-Mead",
or "CG" are used as the \code{solver} argument. One of \code{"Delta", "R", "Rwp"}.
Default = \code{"Rwp"}. See Chipera and Bish (2002) and page 247 of Bish and Post (1989)
for definitions of these functions.}

\item{refs}{A character string of reference pattern IDs or names from the specified library.
The IDs or names supplied must be present within the \code{lib$phases$phase_id} or
\code{lib$phases$phase_name} columns. If missing from the function call then all phases in
the reference library will be used.}

\item{std}{The phase ID (e.g. "QUA.1") to be used as an internal
standard. Must match an ID provided in the \code{refs} parameter.}

\item{force}{An optional string of phase IDs or names specifying which phases should be forced to
remain throughout the automated full pattern summation. The IDs or names supplied must be present
within the \code{lib$phases$phase_id} or \code{lib$phases$phase_name} columns.}

\item{std_conc}{The concentration of the internal standard (if known) in weight percent. If
unknown then omit the argument from the function call or use \code{std_conc = NA} (default),
n which case it will be assumed that all phases sum to 100 percent.}

\item{omit_std}{A logical parameter to be used when the \code{std_conc} argument is defined. When
\code{omit_std = TRUE} the phase concentrations are recomputed to account for the value supplied in
\code{std_conc}. Default \code{= FALSE}.}

\item{normalise}{deprecated. Please use the \code{omit_std} and \code{closed} arguments instead.}

\item{closed}{A logical parameter to be used when the \code{std_conc} argument is defined and
\code{omit_std = TRUE}. When \code{closed = TRUE} the internal standard concentration is removed
and the remaining phase concentrations closed to sum to 100 percent. Default \code{= FALSE}.}

\item{tth_align}{A vector defining the minimum and maximum 2theta values to be used during
alignment (e.g. \code{c(5,65)}). If not defined, then the full range is used.}

\item{align}{The maximum shift that is allowed during initial 2theta
alignment (degrees). Default = 0.1.}

\item{manual_align}{A logical operator denoting whether to optimise the alignment within the
negative/position 2theta range defined in the \code{align} argument, or to use the specified
value of the \code{align} argument for alignment of the sample to the standards. Default
= \code{FALSE}, i.e. alignment is optimised.}

\item{tth_fps}{A vector defining the minimum and maximum 2theta values to be used during
full pattern summation (e.g. \code{c(5,65)}). If not defined, then the full range is used.}

\item{shift}{A single numeric value denoting the maximum (positive or negative) shift,
in degrees 2theta, that is allowed during the shifting of reference patterns. Default = 0.}

\item{remove_trace}{A single numeric value representing the limit for the concentration
of trace phases to be retained, i.e. any mineral with an estimated concentration below
\code{remove_trace} will be omitted. Default = 0.}

\item{weighting}{an optional 2 column data frame specifying the 2theta values in the first
column and a numeric weighting vector in the second column that specifies areas of the pattern
to either emphasise (values > 1) or omit (values = 0) when minimising the objective function
defined in the \code{obj} argument. Use this weighting parameter with caution. The default
is simply a weighting vector where all values are 1, which hence has no effect on the computed
objective function.}

\item{...}{Other parameters passed to methods e.g. \code{fps.powdRlib}}
}
\value{
a powdRfps object with components:
\item{tth}{a vector of the 2theta scale of the fitted data}
\item{fitted}{a vector of the fitted XRPD pattern}
\item{measured}{a vector of the original XRPD measurement (aligned and harmonised)}
\item{residuals}{a vector of the residuals (measured minus fitted)}
\item{phases}{a dataframe of the phases used to produce the fitted pattern and their concentrations}
\item{phases_grouped}{the phases dataframe grouped by phase_name and concentrations summed}
\item{obj}{named vector of the objective parameters summarising the quality of the fit}
\item{weighted_pure_patterns}{a dataframe of reference patterns used to produce the fitted pattern.
All patterns have been weighted according to the coefficients used in the fit}
\item{coefficients}{a named vector of coefficients used to produce the fitted pattern}
\item{inputs}{a list of input arguments used in the function call}
}
\description{
\code{fps} returns estimates of phase concentrations using full pattern
summation of X-ray powder diffraction data.
}
\details{
Applies full pattern summation (Chipera & Bish, 2002, 2013; Eberl, 2003) to an XRPD
measurement to quantify phase concentrations. Requires a \code{powdRlib} library of
reference patterns with reference intensity ratios in order to derive
mineral concentrations. Details provided in Butler and Hillier (2021).
}
\examples{
#Load the minerals library
data(minerals)

# Load the soils data
data(soils)

#Since the reference library is relatively small,
#the whole library can be used at once to get an
#estimate of the phases within each sample.
\dontrun{
fps_sand <-  fps(lib = minerals,
                 smpl = soils$sandstone,
                 refs = minerals$phases$phase_id,
                 std = "QUA.1",
                 align = 0.2)

fps_lime <- fps(lib = minerals,
                smpl = soils$limestone,
                refs = minerals$phases$phase_id,
                std = "QUA.1",
                align = 0.2)

fps_granite <- fps(lib = minerals,
                   smpl = soils$granite,
                   refs = minerals$phases$phase_id,
                   std = "QUA.1",
                   align = 0.2)

#Alternatively run all 3 at once using lapply

fps_soils <- lapply(soils, fps,
                    lib = minerals,
                    std = "QUA.2",
                    refs = minerals$phases$phase_id,
                    align = 0.2)

#Using the rockjock library:

data(rockjock)
data(rockjock_mixtures)

rockjock_1 <- fps(lib = rockjock,
                  smpl = rockjock_mixtures$Mix1,
                  refs = c("ORDERED_MICROCLINE",
                           "LABRADORITE",
                           "KAOLINITE_DRY_BRANCH",
                           "MONTMORILLONITE_WYO",
                           "ILLITE_1M_RM30",
                           "CORUNDUM"),
                  std = "CORUNDUM",
                  align = 0.3)

#Alternatively you can specify the internal standard
#concentration if known:
rockjock_1s <- fps(lib = rockjock,
                 smpl = rockjock_mixtures$Mix1,
                 refs = c("ORDERED_MICROCLINE",
                          "LABRADORITE",
                          "KAOLINITE_DRY_BRANCH",
                          "MONTMORILLONITE_WYO",
                          "ILLITE_1M_RM30",
                          "CORUNDUM"),
                 std = "CORUNDUM",
                 std_conc = 20,
                 align = 0.3)

}
}
\references{
Butler, B. M., Hillier, S., 2021.powdR: An R package for quantitative mineralogy using full pattern
summation of X-ray powder diffraction data. Comp. Geo. 147, 104662. doi:10.1016/j.cageo.2020.104662

Chipera, S.J., Bish, D.L., 2013. Fitting Full X-Ray Diffraction Patterns for Quantitative Analysis:
A Method for Readily Quantifying Crystalline and Disordered Phases. Adv. Mater. Phys. Chem. 03, 47-53.
doi:10.4236/ampc.2013.31A007

Chipera, S.J., Bish, D.L., 2002. FULLPAT: A full-pattern quantitative analysis program for X-ray powder
diffraction using measured and calculated patterns. J. Appl. Crystallogr. 35, 744-749.
doi:10.1107/S0021889802017405

Eberl, D.D., 2003. User's guide to RockJock - A program for determining quantitative mineralogy from
powder X-ray diffraction data. Boulder, CA.
}
