% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minerals}
\alias{minerals}
\title{An example powdRlib reference library}
\format{
A powdRlib object of 3 components
\describe{
\item{xrd}{A dataframe of all the count intensities of all reference patterns. Column names denote the
unique phase ID of each reference pattern}
\item{tth}{A vector of the 2theta scale for all reference patterns in the library}
\item{phases}{A dataframe the phase IDs, names and reference intensity
ratios (RIR)}
}
}
\usage{
minerals
}
\description{
This \code{powdRlib} object, built using the \code{powdRlib} constructor function,
contains a range of measured XRPD data (collected using Cu K-alpha radiation) along with
their reference intensity ratios. The library is designed for simple examples only and
can be used with the \code{soils} data for relatively fast tests of \code{fps} and
\code{afps}.
}
\keyword{datasets}
