% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minerals_phases}
\alias{minerals_phases}
\title{Example phases table for a reference library}
\format{
A 3 column data frame consisting of:
\describe{
\item{phase_id}{A string defining the unique phase IDs that
should match those defined as column names of the minerals table
(e.g. \code{minerals_xrd}).}
\item{phase_name}{A string defining the mineral group that each
reference pattern belongs to.}
\item{rir}{A vector defining the reference intensity ratios
of each reference pattern.}
}
}
\usage{
minerals_phases
}
\description{
A data frame of associated phase information for the \code{minerals_xrd} data.
Together these two data frames can be combined with the \code{powdRlib} constructor
function to create an example reference library (see \code{?powdRlib}). Use the same
layout to create custom reference libraries.
}
\keyword{datasets}
