% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_xrpd_align.R
\name{multi_xrpd_align}
\alias{multi_xrpd_align}
\title{Align and harmonise multiple XRPD patterns}
\usage{
multi_xrpd_align(x, standard, xmin, xmax, xshift)
}
\arguments{
\item{x}{a list XRPD dataframes (column 1 = 2theta, column 2 = counts)}

\item{standard}{a dataframe of the chosen standard that each
sample is aligned to (column 1 = 2theta, column 2 = counts)}

\item{xmin}{the minimum 2theta value used during alignment}

\item{xmax}{the maximum 2theta value used during alignment}

\item{xshift}{the maximum (positive and negative) 2theta shift
that is allowed during alignment}
}
\value{
a \code{multiXY} object of aligned and harmonised XRPD data.
}
\description{
\code{multi_xrpd_align} takes a list of XRPD data and aligns them
relative to a chosen standard. It uses an optimisation routine that
computes a suitable linear shift. After all samples have been aligned,
the function harmonises the data to a single 2theta scale.
}
\examples{
# Load soils xrd data
data(soils)

#Load minerals library
data(minerals)

#Create a standard quartz pattern to align to
quartz <- data.frame(tth = minerals$tth,
                     counts = minerals$xrd$QUA.1)

#Plot the main quartz peak prior to alignment (scale the counts using rng.nm)
plot(soils, wavelength = "Cu",
     xlim = c(26,27),
     normalise = TRUE)

#align data
aligned <- multi_xrpd_align(soils,
                            standard = quartz,
                            xmin = 10,
                            xmax = 60,
                            xshift = 0.2)
#replot data
plot(aligned, wavelength = "Cu",
     xlim = c(26,27),
     normalise = TRUE)

}
