% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powdR.R
\docType{package}
\name{powdR}
\alias{powdR}
\alias{powdR-package}
\title{powdR: A package for XRPD data analysis.}
\description{
The powdR package provides a host of function relevant to XRPD data sets,
including peak alignment and full pattern fitting.
}
