% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powdR.R
\docType{package}
\name{powdR}
\alias{powdR}
\title{powdR: Full Pattern Summation of X-Ray Powder Diffraction Data}
\description{
An implementation of the full pattern summation approach to quantitative mineralogy
from X-ray powder diffraction data (Chipera & Bish, 2002, 2013; Eberl, 2003;
Butler & Hillier 2021).
}
\references{
Butler, B. M., Hillier, S., 2021.powdR: An R package for quantitative mineralogy using full pattern
summation of X-ray powder diffraction data. Comp. Geo. 147, 104662. doi:10.1016/j.cageo.2020.104662

Chipera, S.J., Bish, D.L., 2013. Fitting Full X-Ray Diffraction Patterns for Quantitative Analysis:
A Method for Readily Quantifying Crystalline and Disordered Phases. Adv. Mater. Phys. Chem. 03, 47-53.
doi:10.4236/ampc.2013.31A007

Chipera, S.J., Bish, D.L., 2002. FULLPAT: A full-pattern quantitative analysis program for X-ray powder
diffraction using measured and calculated patterns. J. Appl. Crystallogr. 35, 744-749.
doi:10.1107/S0021889802017405

Eberl, D.D., 2003. User's guide to ROCKJOCK - A program for determining quantitative mineralogy from
powder X-ray diffraction data. Boulder, CA.
}
\author{
Benjamin Butler, The James Hutton Institute, Aberdeen, UK
}
