% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_library.R
\name{powdRlib}
\alias{powdRlib}
\title{Create an XRD reference library}
\usage{
powdRlib(xrd_table, phases_table)
}
\arguments{
\item{xrd_table}{A data frame of the xrd reference patterns and their associated two
theta axis. Must all be on the same two theta scale.}

\item{phases_table}{A data frame of the required data for each reference pattern}
}
\value{
a list with components:
\item{xrd}{a data frame of reference patterns}
\item{tth}{a vector of the two theta measurement intervals}
\item{phases}{a data frame of the id, name and reference
intensity ratio of each reference pattern}
}
\description{
\code{powdRlib} creates a reference library object from two tables of
data. This reference library object is required for the \code{fps} function.
}
\examples{
#load an example xrd_table
data(minerals_xrd)
#load an example phases_table
data(minerals_phases)

#Create a reference library object
xrd_lib <- powdRlib(xrd_table = minerals_xrd,
                    phases_table = minerals_phases)
}
