% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rockjock_regroup}
\alias{rockjock_regroup}
\title{Regrouping structure for the rockjock reference library}
\format{
A data frame with three columns:
\describe{
\item{phase_id}{the phase IDs present in \code{afsis$phases$phase_id}}.
\item{phase_name_grouped}{The phase names that constitute the first
regrouping structure.}
\item{phase_name_grouped2}{The phase names that constitute the second
regrouping structure}
}
}
\usage{
rockjock_regroup
}
\description{
A data frame containing an example re-grouping structure for the \code{rockjock}
reference library, which results in a slightly coarser description of clay
minerals and Fe/Ti-(hydr)oxides in \code{powdRfps} or \code{powdRafps} objects
when used with \code{regroup()}.
}
\keyword{datasets}
