% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.powdRlib}
\alias{subset.powdRlib}
\title{subset a powdRlib object}
\usage{
\method{subset}{powdRlib}(lib, refs, mode, ...)
}
\arguments{
\item{lib}{a powdRlib object}

\item{refs}{a strong of the phase ID's of reference patterns to be subset}

\item{mode}{denotes whether the phase ID's defined in the \code{refs} argument are
retained (\code{"keep"}) or removed (\code{"remove"}).}

\item{...}{other arguments}
}
\description{
\code{subset.powdRlib} is designed to provide an easy way of
subsetting a \code{powdRlib} object by defining the phase ID's
that the user wishes to either keep or remove.
}
\details{
The only mandatory argument is \code{lib}, which must be a powdRlib object.
}
\examples{
#Load the minerals library
data(minerals)

minerals_keep <- subset(minerals,
                        refs = c("QUA.1", "QUA.2"),
                        mode = "keep")

minerals_remove <- subset(minerals,
                          refs = c("QUA.1", "QUA.2"),
                          mode = "remove")
}
