% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd_align.R
\name{xrd.align}
\alias{xrd.align}
\title{Align XRPD data}
\usage{
xrd.align(xrd.sample, xrd.standard, xmin, xmax, xshift)
}
\arguments{
\item{xrd.sample}{a data frame of the XRPD sample to be aligned. The first column
must be 2theta, and the second must be counts}

\item{xrd.standard}{a  data frame of the XRPD sample to be used as the standard. The
first column must be 2theta, and the second must be counts.}

\item{xmin}{The minimum 2theta value to be used during alignment}

\item{xmax}{the maximum 2theta value to be used during alignment}

\item{xshift}{the maximum 2theta shift that can be applied}
}
\value{
a list with components:
\item{shift}{the 2theta shift applied to the sample}
\item{aligned}{a dataframe of the aligned sample (2theta and counts)}
}
\description{
\code{xrd.align} returns a list containing the shift extent (2theta) and a
data frame of the shifted xrpd pattern.
}
\details{
This function applies a linear alignment to an xrpd sample pattern relative to a
chosen standard. The function using brent optimisation to maximise the correlation
between the 2 patterns. Appropriate choice of internal standard is required for
accurate alignment.
}
\examples{
data(soils)
data(minerals)

#Create a quartz standard from the minerals library
quartz <- data.frame(tth = minerals$tth,
                     counts = minerals$xrd$QUA.1)

#Create a soil XRPD pattern to align
soil <- soils$sandstone

#Plot unaligned data. Note that rng.nm is used to normalise count intensities
plot(x = quartz$tth, y = rng.nm(quartz$counts), xlim = c(26, 27), type = "l")
lines(x = soil$tth, y = rng.nm(soil$counts), col = "red")

#Align the soil sample to the quartz
soil_a <- xrd.align(xrd.sample = soil, xrd.standard = quartz, xmin = 10,
                     xmax = 60, xshift = 0.2)
#Replot
plot(x = quartz$tth, y = rng.nm(quartz$counts), xlim = c(26, 27), type = "l")
lines(x = soil_a$aligned$tth, y = rng.nm(soil_a$aligned$counts), col = "red")
}
