% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd_bkg.R
\name{xrd.bkg}
\alias{xrd.bkg}
\title{Fit a background to XRPD data}
\usage{
xrd.bkg(tth, counts, width, res)
}
\arguments{
\item{tth}{the 2theta scale}

\item{counts}{the XRPD counts}

\item{width}{the bin width}

\item{res}{the span of the loess polynomial}
}
\description{
\code{xrd.bkg} returns a vector of background counts
}
\details{
This function bins the data into chunks and takes the minimum counts
from each chunk. A loess polynomial is then fitted between the minimum
values and the polynomial predicted onto the original 2theta scale.
}
